fixed the migration of the database
We had too many FK keys in the account class which meant that EF core could not figure out by itself what should be paired with what.
