+++ MandelsBankenConsole/Models/User.cs
@@ -1,6 +1,6 @@
 ﻿namespace MandelsBankenConsole.Models
 {
-    internal class User
+    public class User
     {
         public int Id { get; set; }
         public string CustomerName { get; set; }
@@ -8,7 +8,12 @@
         public string Pin { get; set; }
 
         // Navigation property
-        public virtual ICollection<Account> Accounts { get; set; }
+        public virtual ICollection<Account> Accounts { get; set; } = new List<Account>();
     }
 }
+
+// A user can have multiple accounts, each account has refernce to a currency,
+// all have a collection of transactions thats unique to that specifec account and all the transactions has a reference to the account it belongs to.
+
+
